Removed SQLite & OnlineOfflineHandler
Only online handling will be nessesairy in the model. The offline
database and the online offline handler should be handled by the
implementing application. (Specific for this project: The
android.database.sqlite.SQLiteDatabase will be responsible for handling
the offline database on the application. As well as the OnlineOffline
transaction between server and phone.